% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sheet.R
\name{add_sheet}
\alias{add_sheet}
\title{Add an Excel worksheet through \code{openxlsx} if it doesn't already exist}
\usage{
add_sheet(wb, sheet_name)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet}

\item{sheet_name}{The desired worksheet name}
}
\value{
Worksheet added (if required) to the wb object.
}
\description{
Checks if the sheet with the specified name exists, and creates it if needed.
}
\examples{
\dontrun{
# To add
}
}
