% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_converged.R
\name{check_converged}
\alias{check_converged}
\title{Check whether survival models extrapolations converge for survival analysis}
\usage{
check_converged(formula, data, dists)
}
\arguments{
\item{formula}{a formula specifying the model to be used, in the form of
\code{Surv(time,event)~strata[+covariates]} in /code{flexsurv} terms.}

\item{data}{A data frame containing the data to be used for the analysis.
This must contain data for the \code{time} and \code{event} variables.}

\item{dists}{Name or vector of names for the distribution(s) to be used.}
}
\value{
A vector of distribution names that have converged.
}
\description{
This function checks whether survival model extrapolations converge by
fitting the models using the \code{\link[survHE]{fit.models}} function and
checking for convergence. Extrapolations that do not converge are excluded
from the output.
}
\examples{
\dontrun{
library(survival)

check_converged(
  formula = Surv(time, status) ~ as.factor(sex),
  data = lung[1:7,],
  dists = c("exponential", "weibull", "gengamma", "lognormal")
)
}

}
