% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_converged_splines.R
\name{check_converged_splines}
\alias{check_converged_splines}
\title{Check whether spline models converge}
\usage{
check_converged_splines(formula, data, dists)
}
\arguments{
\item{formula}{A formula specifying the model to be used, in the form
\code{Surv(time, event) ~ strata[+covariates]} in flexsurv terms.}

\item{data}{A data frame containing the data to be used for the analysis.
This must contain data for the \code{time} and \code{event} variables.}

\item{dists}{A data frame with columns \code{knots} and \code{scale}
specifying the number of knots to be used and the type of spline model to be
executed respectively. The possible models are: Proportional hazards model
(\code{scale = "hazard"}), probit model (\code{scale = "normal"}),
proportional odds model (\code{scale = "odds"}).}
}
\value{
A modified version of the \code{dists} input dataframe after
removing the models that did not converge.
}
\description{
This function checks whether spline models specified by different knot
configurations converge.
}
\examples{
\dontrun{
library(survival)

formula <- Surv(time, status) ~ as.factor(sex)
dists <- data.frame(knots = c(3, 5),
                    scale = rep("normal", 2))
check_converged_splines(formula, data = lung[1:9,], dists)
}

}
