% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{easy_bc}
\alias{easy_bc}
\title{Formatted Copy of \link[flexsurv:bc]{flexsurv::bc}}
\format{
The data set contains the following variables:

\describe{
\item{censrec}{0 = Censored, 1 = Dead}
\item{rectime}{Time of censoring or death in days}
\item{group}{Prognostic group: Good, Medium, or Poor}
\item{recyrs}{Time of censoring or death in years}
}
}
\source{
\link[flexsurv:bc]{flexsurv::bc}
}
\usage{
easy_bc
}
\description{
This is a copy of the bc data set exported by the flexsurv
package. This data set, however, has column labels assigned.
}
\keyword{datasets}
