% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{easy_lung}
\alias{easy_lung}
\title{Formatted Copy of \code{survival::lung}}
\format{
The data set contains the following variables:

\describe{
\item{inst}{Institution code}
\item{time}{Survival time, months}
\item{status}{Censoring status, 1 = censored, 2 = dead}
\item{age}{Age}
\item{sex}{Sex, 1 = Male, 2 = Female}
\item{ph.ecog}{ECOG Performance Status (Physician)}
\item{ph.karno}{Karnofsky performance score (Physician)}
\item{pat.karno}{Karnofsky performance score (Patient)}
\item{meal.cal}{Calories consumed}
\item{wt.loss}{Weight loss, lbs}
}
}
\source{
\code{survival::lung}
}
\usage{
easy_lung
}
\description{
This is a copy of the lung data set exported by the survival
package. This data set, however, has column labels assigned and time in
months.
}
\keyword{datasets}
