% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_models}
\alias{fit_models}
\title{Fit Survival Models}
\usage{
fit_models(
  data,
  time,
  event,
  group = NULL,
  group_as_covariate = FALSE,
  other_covariates = NULL,
  dists = c("exp", "gamma", "gengamma", "gompertz", "llogis", "lnorm", "weibull"),
  eval_time = NULL,
  engine = "flexsurv",
  k = c(1, 2, 3),
  scale = c("hazard"),
  include_ci = FALSE
)
}
\arguments{
\item{data}{A data frame containing the survival data.}

\item{time}{The name of the column in \code{data} containing the
time-to-event information.}

\item{event}{The name of the column in \code{data} indicating whether the
event of interest occurred.}

\item{group}{(Optional) The name of the column in \code{data} defining the
grouping variable.}

\item{group_as_covariate}{(Optional) Logical indicating whether the grouping
variable should be treated as a covariate.}

\item{other_covariates}{(Optional) A character vector specifying the names of
other covariates to be included in the model (not the group variable).}

\item{dists}{(Optional) A character vector specifying the distribution(s) to
be fitted.

For flexsurv-based engines, options are "exp", "gamma", "genf", "genf.orig",
"gengamma", "gengamma.orig", "gompertz", "llogis", "lnorm", "lognormal",
"weibull", and "weibullPH".

For the survival engine, options are "exponential", "extreme", "gaussian",
"loggaussian", "logistic", "lognormal", "rayleigh", "weibull".

Default is \code{c("exp", "gamma", "gengamma", "gompertz",
"llogis", "lnorm", "weibull")} which applies to flexsurv-related engines.}

\item{eval_time}{(Optional) A vector of evaluation time points for generating
predictions. Default is \code{NULL}, which if left as NULL, generates a
sequence from 0 to 2.5 times the maximum observed time.}

\item{engine}{(Optional) The survival analysis engine to be used.
Options are "flexsurv", "flexsurvcure", "flexsurvspline", and "survival".
Default is "flexsurv".}

\item{k}{(Optional) A numeric vector specifying the number of knots for
spline-based models. Default is \code{c(1, 2, 3)} to test different numbers.}

\item{scale}{(Optional) A character vector specifying the scale parameter(s)
for spline-based models. Options are "hazard", "odds", and "normal".
Default is \code{"hazard"}.}

\item{include_ci}{(Optional) Logical indicating whether to include confidence
intervals in the predictions. Default is \code{FALSE}.}
}
\value{
A list containing various outputs including model distributions,
parameters, predictions, plots, and summary statistics.
}
\description{
Fits survival models to the provided data using the specified engine and
returns various outputs including model parameters, predictions, and plots.
}
\examples{
\dontrun{

output_separate <- fit_models(
  data = easysurv::easy_bc,
  time = "recyrs",
  event = "censrec",
  group = "group",
  group_as_covariate = FALSE)

}
}
