% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit_averages.R
\name{get_fit_averages}
\alias{get_fit_averages}
\title{Obtain median, mean, and restricted mean survival times.}
\usage{
get_fit_averages(mod)
}
\arguments{
\item{mod}{A specific model of a \code{fit.models} object obtained from
the \code{\link[survHE]{fit.models}} function.}
}
\value{
A data frame containing the comparison of goodness-of-fit measures
for the models. The data frame has columns for the model names, AIC values,
BIC values, AIC ranks, and BIC ranks.
}
\description{
This function obtains the median, mean, and restricted mean survival times
stored in a \code{flexsurvreg} or \code{\link[survHE]{fit.models}} object.
}
\examples{
\dontrun{
library(survHE)
library(flexsurv)

# Example usage with a fit.models object
fit <- fit.models(Surv(time, status) ~ as.factor(sex),
  data = lung,
  distr = c("exponential", "weibull")
)
lapply(fit$models, get_fit_averages)

# Example using a flexsurv object
fit <- flexsurvreg(Surv(time, status) ~ as.factor(sex),
  data = lung,
  dist = "exponential"
)
get_fit_averages(fit)
}
}
