% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit_coef.R
\name{get_fit_coef}
\alias{get_fit_coef}
\title{Get parameter estimates from a fit.models object}
\usage{
get_fit_coef(mod, method = "mle")
}
\arguments{
\item{mod}{An object of class \code{flexsurv} obtained from
the\code{\link[survHE]{fit.models}}, \code{\link[flexsurv]{flexsurvreg}}
or \code{\link[flexsurvcure]{flexsurvcure}} functions.}

\item{method}{The method used to estimate the parameters. The default is
\code{'mle'} which corresponds to maximum likelihood estimation.
Alternatively, if \code{'hmc'} is specified, the function assumes
the model was fitted using Hamiltonian Monte Carlo (HMC) via the
\code{`rstan`} package. In this case, the parameter estimates are obtained
from the HMC model. This is not imported intentionally.}
}
\value{
A data frame containing the parameter estimates from the
fitted models. The data frame has two columns: \code{parameter} and
\code{Coef}. The \code{parameter} column lists the names of the model
parameters, and the \code{Coef} column contains their corresponding
estimated values.
}
\description{
This function retrieves the parameter estimates from a
\code{\link[flexsurv]{flexsurvreg}} or
\code{\link[flexsurvcure]{flexsurvcure}} object,
which are listed in the output of the \code{\link[survHE]{fit.models}}
function under \code{$models}.
The flexsurv object can alternately be created using
\code{\link[flexsurv]{flexsurvreg}} or
\code{\link[flexsurvcure]{flexsurvcure}}.
}
\examples{
\dontrun{
library(survHE)
library(flexsurv)

# Example usage with a fit.models object
fit <- fit.models(Surv(time, status) ~ as.factor(sex),
  data = lung,
  distr = c("exponential", "weibull")
)
lapply(fit, get_fit_coef)

# Example using a flexsurv object
fit <- flexsurvreg(Surv(time, status) ~ as.factor(sex),
  data = lung,
  dist = "exponential"
)
get_fit_coef(fit)

# Example usage with a fit.models object fitted using HMC
fit_hmc <- fit.models(Surv(time, status) ~ as.factor(sex),
  data = lung,
  distr = c("exponential", "weibull"), method = "hmc"
)
lapply(fit_hmc, get_fit_coef)
}
}
