% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit_comparison.R
\name{get_fit_comparison}
\alias{get_fit_comparison}
\title{Compare statistical goodness-of-fit between models in a fit.models object}
\usage{
get_fit_comparison(mod)
}
\arguments{
\item{mod}{An object of class \code{fit.models} obtained from the
\code{\link[survHE]{fit.models}} function.}
}
\value{
A data frame containing the comparison of goodness-of-fit measures
for the models. The data frame has columns for the model names, AIC values,
BIC values, AIC ranks, and BIC ranks.
}
\description{
This function compares the statistical goodness-of-fit between the models
stored in a \code{\link[survHE]{fit.models}} object. It retrieves and
organizes AIC and BIC and ranks them accordingly.
}
\examples{
\dontrun{
library(survHE)

fit <- fit.models(Surv(time, status) ~ as.factor(sex),
  data = lung,
  distr = c("exponential", "weibull")
)
fit_comparison <- get_fit_comparison(fit)
print(fit_comparison)
}

}
