% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_median_FU.R
\name{get_median_FU}
\alias{get_median_FU}
\title{Calculate the median follow-up time (or median time to censoring).}
\usage{
get_median_FU(data, time, event, group = NULL)
}
\arguments{
\item{data}{A tibble or data frame containing the survival data with
columns for \code{time}, \code{event} and \code{strata}.}

\item{time}{The name of the time variable in data}

\item{event}{The name of the event variable in data}

\item{group}{The name of the group variable in data}
}
\description{
Reverses the censoring/event variable to derive median follow-up time.
}
\examples{
\dontrun{
# Load required packages
library(survival)

input_data <- survival::lung

surv_data <- tibble(
  time = input_data$time,
  event = input_data$status - 1,
  strata = as.factor(input_data$sex)
)

median_follow_up <- get_median_FU(
  data = surv_data,
  time = "time",
  event = "event",
  strata = "strata"
)
}
}
