% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_table.R
\name{get_results_table}
\alias{get_results_table}
\title{Generate a parameter table with the vcov matrix from a fit.models object}
\usage{
get_results_table(mod)
}
\arguments{
\item{mod}{The "$models" element on an object of class \code{fit.models}
(i.e. \code{fit.models.object[["models"]]}) obtained from the
\code{\link[survHE]{fit.models}} function. Ensure that, in the formula used
to make \code{mod} (\code{Surv(time,event)~strata[+covariates]}),
\code{class(strata) == "factor"}.}
}
\value{
A data frame representing the parameter table with estimated
coefficients and the named variance-covariance elements. The table contains
columns for the model labels, parameter names, estimated coefficients,
and the named variance-covariance elements.
}
\description{
This function generates a parameter table along with the variance-covariance
matrix obtained from the "models" element of a
\code{\link[survHE]{fit.models}} object. The parameter table includes
information such as the model labels, parameter names, estimated
coefficients, and the named variance-covariance elements. The function also
assigns appropriate labels to the covariance markers for better
interpretation. This table is useful for summarizing the results of fitted
survival models and can be exported or used for further analysis.
}
\examples{
\dontrun{
library(survHE)

fit <- fit.models(Surv(time, status) ~ as.factor(sex),
  data = lung,
  distr = c("exponential", "weibull")
)
results_table <- get_results_table(fit$models)
print(results_table)
}

}
