% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_table_cure.R
\name{get_results_table_cure}
\alias{get_results_table_cure}
\title{Generate a parameter table with vcov matrix the object generated by
fit.models.cure}
\usage{
get_results_table_cure(mod)
}
\arguments{
\item{mod}{The "$models" element of the list object created by a single-arm
(no treatment effect) \code{\link{fit.models.cure}} function.}
}
\value{
A data frame with the parameter table and variance-covariance matrix
from the cure models.
}
\description{
This function generates a parameter table along with the variance-covariance
matrix obtained from a list of cure models. The parameter table includes
information such as the model labels, parameter estimates, and associated
variance-covariance matrices.
}
\examples{
\dontrun{
library(survHE)
fit <- fit.models.cure(Surv(time, status) ~ 1,
  data = lung, dist = c("exponential", "weibull")
)
get_results_table_cure(fit$models)
}

}
