% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_FU.R
\name{median_FU}
\alias{median_FU}
\title{Calculate the median follow-up time (or median time to censoring).}
\usage{
median_FU(data, time, event, strata)
}
\arguments{
\item{data}{A tibble or data frame containing the survival data with
columns for \code{time}, \code{event} and \code{strata}.}

\item{time}{The name of the time variable in data}

\item{event}{The name of the event variable in data}

\item{strata}{The name of the strata variable in data}
}
\description{
Reverses the censoring/event variable to derive median follow-up time.
}
\examples{
\dontrun{
# Load required packages
library(survival)

input_data <- survival::lung

surv_data <- tibble(
  time = input_data$time,
  event = input_data$status - 1,
  strata = as.factor(input_data$sex)
)

median_follow_up <- median_FU(
  data = surv_data,
  time = "time",
  event = "event",
  strata = "strata"
)
}
}
