% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fits}
\alias{plot_fits}
\title{Plot Survival Fits}
\usage{
plot_fits(data)
}
\arguments{
\item{data}{A data frame containing the survival predictions.}
}
\value{
A ggplot2 object.
}
\description{
Plots the survival fits generated by the \code{fit_models} function.
}
