% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hazards}
\alias{plot_hazards}
\title{Plot Smoothed Hazards}
\usage{
plot_hazards(
  pred_data,
  obs_data,
  legend_label = "Model",
  title = NULL,
  subtitle = NULL,
  plot_theme = theme_easysurv()
)
}
\arguments{
\item{pred_data}{A data frame containing the hazard predictions.}

\item{obs_data}{A data frame containing the observed hazards.}

\item{legend_label}{Label for the legend. Default is "Model".}

\item{title}{Title of the plot. Default is \code{NULL}.}

\item{subtitle}{Subtitle of the plot. Default is \code{NULL}.}

\item{plot_theme}{ggplot2 theme for the plot. Default is
\code{theme_easysurv()}.}
}
\value{
A ggplot2 object.
}
\description{
Plots observed hazards and the hazards predicted by different models
generated by the \code{fit_models} function.
}
