% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_KM.R
\name{plot_KM}
\alias{plot_KM}
\title{Plots a Kaplan-Meier Survival Curve using \code{ggsurvplot} with
Custom Defaults}
\usage{
plot_KM(
  fit,
  title = NULL,
  subtitle = NULL,
  legend.labs = NULL,
  legend.title = ggplot2::element_blank(),
  legend.position = "top",
  conf.int = TRUE,
  conf.int.alpha = 0.3,
  conf.int.style = "ribbon",
  tables.col = "black",
  risk.table = TRUE,
  risk.table.pos = "out",
  risk.table.fontsize = 4,
  risk.table.title = "Number at risk",
  cumevents = FALSE,
  cumcensor = FALSE,
  surv.median.line = "hv",
  axes.offset = FALSE,
  font.family = "Roboto Condensed",
  plot.theme = theme_easysurv(),
  risk.table.theme = theme_easysurv(grid = FALSE),
  use_plotly = FALSE,
  fun = NULL,
  surv.scale = "percent",
  xlim = c(0, max(fit$time)),
  ...
)
}
\arguments{
\item{fit}{A \code{\link[survival]{survfit}} object or list of
\code{\link[survival]{survfit}} objects.}

\item{title}{Optional. A title for the KM plot. Default is NULL}

\item{subtitle}{Optional. A subtitle for the KM plot. Default is NULL}

\item{legend.labs}{character vector specifying legend labels. Used to replace
the names of the strata from the fit. Should be given in the same order as
those strata.}

\item{legend.title}{The title of the legend.}

\item{legend.position}{The position of the legend in the plot.
Allowed values are \code{"left"}, \code{"right"}, \code{"bottom"},
\code{"top"}, or \code{"none"}.}

\item{conf.int}{Logical value indicating whether to display the confidence
interval.}

\item{conf.int.alpha}{The transparency level of the confidence interval fill
color. Value should be in the range of <0, 1>, where 0 is fully transparent
and 1 is fully opaque.}

\item{conf.int.style}{The style of the confidence interval.
Allowed values are \code{"ribbon"} or \code{"step"}.}

\item{tables.col}{Colour to be used for all tables under the main plot.
Default is "black". Can be "strata".}

\item{risk.table}{Logical value indicating whether to show the
"number at risk" table.}

\item{risk.table.pos}{The position of the risk table relative to
the main plot.
Allowed values are \code{"out"} (outside) or \code{"in"} (inside).}

\item{risk.table.fontsize}{The font size for the risk table.}

\item{risk.table.title}{The title for the risk table.}

\item{cumevents}{Logical to include cumulative events table. Default is FALSE}

\item{cumcensor}{Logical to include cumulative censoring table.
Default is FALSE}

\item{surv.median.line}{Specifies whether and how to draw a horizontal
or vertical line at the median survival time.
Allowed values are \code{"none"}, \code{"hv"} (both horizontal and vertical),
\code{"h"} (horizontal), or \code{"v"} (vertical).}

\item{axes.offset}{Logical value indicating whether to set the plot axes to
start at the origin. Default is FALSE. When FALSE, recommend
that xlim is set to something like c(0, max(fit$time)) to not cut off
the end.}

\item{font.family}{The font for the chart and risk table. Default is
"Roboto Condensed".}

\item{plot.theme}{function, ggplot2 theme name. Default value is
\code{theme_easysurv()}. Allowed values include ggplot2 official themes:
see \code{\link[ggplot2]{theme}}.}

\item{risk.table.theme}{function, ggplot2 theme name. Default value is
\code{theme_easysurv(grid = FALSE)}. Allowed values include ggplot2
official themes:see \code{\link[ggplot2]{theme}}.}

\item{use_plotly}{Optional. Whether to return a \code{plotly} output
for the KM plot. Default is FALSE.}

\item{fun}{Survival function. Can take on "survival", "event", "cumhaz",
"cloglog"}

\item{surv.scale}{Scale transformation of survival curves.
ggsurvplot allows "default" or "percent". When FUN is not specified,
the easysurv default is "percent".}

\item{xlim}{Two numeric values, specifying the left/lower limit and the
right/upper limit of the scale.}

\item{...}{Additional arguments to be passed to the
\code{\link[survminer]{ggsurvplot}} function.}
}
\value{
A list containing the plot object and the risk table object.
}
\description{
This function generates a Kaplan-Meier survival curve plot using the
\code{\link[survminer]{ggsurvplot}} function.
It allows for customization of various plot elements while providing
sensible defaults that align with the easysurv theme.
}
\examples{
\dontrun{
library(survival)
fit <- survfit(Surv(time, status) ~ as.factor(sex), data = lung)
plot_KM(fit,
legend.title = "Sex", legend.position = "bottom", conf.int = TRUE)
}

}
