% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_schoenfeld}
\alias{plot_schoenfeld}
\title{Plot Schoenfeld Residuals}
\usage{
plot_schoenfeld(
  residuals,
  hline = TRUE,
  sline = TRUE,
  sline.se = TRUE,
  hline.col = "red",
  hline.size = 1,
  hline.alpha = 1,
  hline.yintercept = 0,
  hline.lty = "dashed",
  sline.col = "blue",
  sline.size = 1,
  sline.alpha = 0.3,
  sline.lty = "dashed",
  point.col = "black",
  point.size = 1,
  point.shape = 19,
  point.alpha = 1,
  plot_theme = ggplot2::theme_bw()
)
}
\arguments{
\item{residuals}{A data frame containing the Schoenfeld residuals, typically
with columns \code{time}, \code{residual}, \code{transform}, and \code{variable}.}

\item{hline}{Logical. If \code{TRUE}, a horizontal line is added to the plot.
Default is \code{TRUE}.}

\item{sline}{Logical. If \code{TRUE}, a smooth line is added to the plot.
Default is \code{TRUE}.}

\item{sline.se}{Logical. If \code{TRUE}, confidence intervals are displayed around
the smooth line. Default is \code{TRUE}.}

\item{hline.col}{Color of the horizontal line. Default is \code{"red"}.}

\item{hline.size}{Line width of the horizontal line. Default is \code{1}.}

\item{hline.alpha}{Transparency of the horizontal line. Default is \code{1}.}

\item{hline.yintercept}{Y-intercept for the horizontal line. Default is \code{0}.}

\item{hline.lty}{Line type for the horizontal line. Default is \code{"dashed"}.}

\item{sline.col}{Color of the smooth line. Default is \code{"blue"}.}

\item{sline.size}{Line width of the smooth line. Default is \code{1}.}

\item{sline.alpha}{Transparency of the smooth line. Default is \code{0.3}.}

\item{sline.lty}{Line type for the smooth line. Default is \code{"dashed"}.}

\item{point.col}{Color of the points representing residuals. Default is
\code{"black"}.}

\item{point.size}{Size of the points representing residuals. Default is \code{1}.}

\item{point.shape}{Shape of the points representing residuals. Default is
\code{19}.}

\item{point.alpha}{Transparency of the points representing residuals. Default
is \code{1}.}

\item{plot_theme}{A ggplot2 theme for the plot. Default is
\code{ggplot2::theme_bw()}.}
}
\value{
A ggplot object representing the plot of Schoenfeld residuals.
}
\description{
This function creates a visual representation of Schoenfeld residuals from a
Cox proportional hazards model.
It allows for customization of the plot, including the addition of horizontal
and smoothed lines, and styling of points and plot elements.
}
