% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smoothed_hazards.R
\name{plot_smoothed_hazards}
\alias{plot_smoothed_hazards}
\title{Plot smoothed hazards of survival data}
\usage{
plot_smoothed_hazards(
  data,
  time,
  event,
  fits,
  t,
  group = 1,
  title = "",
  subtitle = "",
  ylab = "Hazard",
  xlab = "Time",
  ylimit = NULL,
  font.family = "Roboto Condensed",
  plot.theme = theme_easysurv(),
  use_plotly = FALSE
)
}
\arguments{
\item{data}{The survival data as a data frame with columns
\code{time} and \code{event}.}

\item{time}{The name of the time variable in data.}

\item{event}{The name of the event variable in data.}

\item{fits}{The \code{\link[survHE]{fit.models}} object containing
survival models}

\item{t}{A vector of time points at which to calculate the hazards.}

\item{group}{The group number of the strata}

\item{title}{The title of the plot.}

\item{subtitle}{The subtitle of the plot.}

\item{ylab}{The label for the y-axis.}

\item{xlab}{The label for the x-axis.}

\item{ylimit}{The upper limit of the y-axis to adjust the scale of the plot.}

\item{font.family}{The name of the font for the plot. Default is
"Roboto Condensed".}

\item{plot.theme}{ggplot2 theme for the plot. Default is
\code{\link{theme_easysurv}()}.}

\item{use_plotly}{Optional. Whether to return a \code{plotly} output
for the plots. Default is FALSE.}
}
\value{
A ggplot object representing the plot of hazards.
}
\description{
This function generates a plot showing the hazards of the underlying
survival data, along with the hazards
predicted by different models from a \code{\link[survHE]{fit.models}} object.
It provides a visual comparison between observed hazards
and the hazards predicted by different models.
}
\examples{
\dontrun{
library(survival)

example <- lung |>
  mutate(event = status) |>
  select(time, event, sex)

models <- survHE::fit.models(
  formula = Surv(time, event) ~ as.factor(sex),
  data = example,
  distr = c("exp", "weibull"),
  method = "mle"
)

plot_smoothed_hazards(
  data = example,
  time = "time",
  event = "event",
  fits = models,
  t = 0:2000,
  title = "Smoothed hazards"
)
}

}
