% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_surv}
\alias{plot_surv}
\title{Plot Survival Fits}
\usage{
plot_surv(
  pred_data,
  km_data,
  km_include = TRUE,
  legend_label = "Model",
  title = NULL,
  subtitle = NULL,
  plot_theme = theme_easysurv()
)
}
\arguments{
\item{pred_data}{A data frame containing the survival predictions.}

\item{km_data}{A data frame containing the Kaplan-Meier survival data.}

\item{km_include}{Logical value indicating whether to include lines for
Kaplan-Meier survival data. Default is \code{TRUE}.}

\item{legend_label}{Label for the legend. Default is "Model".}

\item{title}{Title of the plot. Default is \code{NULL}.}

\item{subtitle}{Subtitle of the plot. Default is \code{NULL}.}

\item{plot_theme}{ggplot2 theme for the plot. Default is
\code{theme_easysurv()}.}
}
\value{
A ggplot2 object.
}
\description{
Primarily a helper function for \code{predict_and_plot}.
Plots the survival fits generated by the \code{fit_models} function.
}
