% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fits.R
\name{predict_fits}
\alias{predict_fits}
\title{Generate a data frame of predicted survival from a fit.models or
fit.models.cure object}
\usage{
predict_fits(fits, t, group = 1, ci = TRUE)
}
\arguments{
\item{fits}{An object obtained from \code{\link[survHE]{fit.models}} or
\code{\link{fit.models.cure}}.}

\item{t}{A vector of times at which to calculate predicted survival,
e.g., \code{seq(from = 0, to = 1000, length.out = 100)}}

\item{group}{In cases of multiple stratas within one
\code{\link[survHE]{fit.models}} object, the index of the "strata" of
interest (e.g., In data frames with two treatments, \code{group = "1"} or
\code{group = "2"})}

\item{ci}{Logical indicating whether to include confidence intervals in the
output.}
}
\value{
A data frame with predicted survival values for each model at the
specified times.
}
\description{
This function generates a data frame of predicted survival values based on a
\code{\link[survHE]{fit.models}} or \code{\link{fit.models.cure}} object.
It takes a set of times and the group in the data, and returns a data frame
with predicted survival values for each model in the input object at the
specified times.
}
\examples{
\dontrun{
library(survHE)

# Single-arm
fits <- fit.models(Surv(time, status) ~ 1,
  data = lung,
  dist = c("exponential", "weibull")
)
t <- seq(from = 0, to = 1000, length.out = 100)
predict_fits(fits, t, group = 1)

# Double-arm
fits <- fit.models(Surv(time, status) ~ as.factor(sex),
  data = lung,
  dist = c("exponential", "weibull")
)
t <- seq(from = 0, to = 1000, length.out = 100)
predict_fits(fits, t, group = 1)
predict_fits(fits, t, group = 2)
}

}
