% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{print.fit_models}
\alias{print.fit_models}
\title{Print methods for \code{fit_models}}
\usage{
\method{print}{fit_models}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fit_models}}

\item{...}{Additional arguments}
}
\value{
A print summary of the \code{fit_models} object.
}
\description{
Print methods for \code{fit_models}
}
\examples{
models <- fit_models(
  data = easysurv::easy_bc,
  time = "recyrs",
  event = "censrec",
  predict_by = "group",
  covariates = "group"
)

models
}
