% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_km.R
\name{print.get_km}
\alias{print.get_km}
\title{Print methods for \code{get_km()}}
\usage{
\method{print}{get_km}(x, ...)
}
\arguments{
\item{x}{An object of class \code{get_km}}

\item{...}{Additional arguments}
}
\value{
The summary of the Kaplan-Meier estimates, printed via the console.
}
\description{
Print methods for \code{get_km()}
}
\examples{
km_results <- get_km(
  data = easysurv::easy_bc,
  time = "recyrs",
  event = "censrec",
  group = "group",
  risktable_symbols = FALSE
)

print(km_results)
}
