% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_surv_data.R
\name{print.inspect_surv_data}
\alias{print.inspect_surv_data}
\title{Print methods for \code{inspect_surv_data}}
\usage{
\method{print}{inspect_surv_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{inspect_surv_data}}

\item{...}{Additional arguments}
}
\value{
A print summary of the \code{inspect_surv_data} object.
}
\description{
Print methods for \code{inspect_surv_data}
}
