% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_and_plot.R
\name{print.pred_plot}
\alias{print.pred_plot}
\title{Print methods for \code{pred_plot}}
\usage{
\method{print}{pred_plot}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pred_plot}}

\item{...}{Additional arguments}
}
\description{
Print methods for \code{pred_plot}
}
