% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{print.predict_and_plot}
\alias{print.predict_and_plot}
\title{Print methods for \code{predict_and_plot()}}
\usage{
\method{print}{predict_and_plot}(x, ...)
}
\arguments{
\item{x}{An object of class \code{predict_and_plot}}

\item{...}{Additional arguments}
}
\value{
A print summary of the \code{predict_and_plot} object.
}
\description{
Print methods for \code{predict_and_plot()}
}
\examples{
models <- fit_models(
  data = easysurv::easy_bc,
  time = "recyrs",
  event = "censrec",
  predict_by = "group"
)

predict_and_plot(models)
}
