% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ph.R
\name{print.test_ph}
\alias{print.test_ph}
\title{Print methods for \code{test_ph()}}
\usage{
\method{print}{test_ph}(x, ...)
}
\arguments{
\item{x}{An object of class \code{test_ph}}

\item{...}{Additional arguments}
}
\value{
A print summary of the \code{test_ph} object.
}
\description{
Print methods for \code{test_ph()}
}
\examples{
ph_results <- test_ph(
  data = easysurv::easy_bc,
  time = "recyrs",
  event = "censrec",
  group = "group"
)

ph_results
}
