% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_quick_KM.R
\name{print.quick_KM}
\alias{print.quick_KM}
\title{Print quick_KM objects.}
\usage{
\method{print}{quick_KM}(x, ...)
}
\arguments{
\item{x}{a quick_KM object}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Print quick_KM objects.
}
