% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.quick_PH}
\alias{print.quick_PH}
\title{Print quick_PH objects.}
\usage{
\method{print}{quick_PH}(x, ...)
}
\arguments{
\item{x}{a quick_PH object}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Print quick_PH objects.
}
