% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_KM.R
\name{quick_KM}
\alias{quick_KM}
\title{Generate Kaplan-Meier (KM) plot and related KM info.}
\usage{
quick_KM(
  data,
  time,
  event,
  strata,
  strata_labels = NULL,
  title = NULL,
  subtitle = NULL,
  add_time_0 = TRUE,
  add_interactive_plot = FALSE,
  font.family = "Roboto Condensed",
  ...
)
}
\arguments{
\item{data}{A tibble or data frame containing the survival data with
columns for \code{time}, \code{event} and \code{strata}.}

\item{time}{The name of the time variable in data}

\item{event}{The name of the event variable in data}

\item{strata}{The name of the strata variable in data}

\item{strata_labels}{Optional. A character vector containing the names of
the strata (default is NULL). Provide in a consistent order with
\code{levels(as.factor(data$strata))}.}

\item{title}{Optional. A title for the KM plot. Default is NULL}

\item{subtitle}{Optional. A subtitle for the KM plot. Default is NULL}

\item{add_time_0}{Optional. Uses survival::survfit0 to add the point for
a starting time (time 0) to a survfit object's elements.
This is useful for plotting. Default is TRUE.}

\item{add_interactive_plot}{Optional. Whether to include a \code{plotly} output
of the KM plot. Default is FALSE.}

\item{font.family}{The name of the font for the KM plot. Default is
"Roboto Condensed".}

\item{...}{Ellipses to pass further arguments to plot_KM.}
}
\value{
An object of class \code{quick_KM} with components:
\item{KM_all}{\code{survfit} model output based on
\code{survival::survfit(survival::Surv(time, event) ~ as.factor(strata)}}
\item{KM_indiv}{\code{survfit} model output based on
\code{survival::survfit(survival::Surv(time, event) ~ 1} for each strata}
\item{KM_stepped}{A table summarizing KM survival over time, presented in a
'stepped' manner to support plotting in Excel.}
\item{KM_plot}{A KM plot in the \code{easysurv} theme.}
\item{KM_summary}{Summary information for each strata. Includes sample
sizes, number of events, restricted mean survival, median, and confidence
intervals}
}
\description{
Returns a \code{easysurv}-themed Kaplan-Meier (KM) plot and related KM info.
}
\examples{
\dontrun{
# Load required packages
library(survival)

input_data <- survival::lung

surv_data <- tibble(
  time = input_data$time,
  event = input_data$status - 1,
  strata = as.factor(input_data$sex)
)

KM_check <- quick_KM(
  data = surv_data,
  time = "time",
  event = "event",
  strata = "strata"
)
}
}
