% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_PH.R
\name{quick_PH}
\alias{quick_PH}
\title{Assess the proportional hazards assumptions with plots and statistical tests.}
\usage{
quick_PH(
  data,
  time,
  event,
  strata,
  plot.theme = theme_easysurv(),
  risk.table.theme = theme_easysurv(grid = FALSE),
  strata_labels = NULL,
  subtitle = NULL,
  font.family = "Roboto Condensed",
  ...
)
}
\arguments{
\item{data}{A tibble or dataframe containing the survival data with
columns \code{time}, \code{event} and \code{strata}.}

\item{time}{The name of the time variable in data}

\item{event}{The name of the event variable in data}

\item{strata}{The name of the strata variable in data}

\item{plot.theme}{ggplot2 theme for cloglog and schoenfeld plots.
Default is theme_easysurv()}

\item{risk.table.theme}{ggplot2 theme for the risk table of clogclog plot.}

\item{strata_labels}{Optional. A character vector containing the names of
the stratas (default is NULL). Provide in a consistent order with
\code{levels(as.factor(data$strata))}.}

\item{subtitle}{Optional. A subtitle for all plots (default is NULL).
Consider using the endpoint name.}

\item{font.family}{Font for the plot. Default is Roboto Condensed.}

\item{...}{Optional arguments for the cloglogplot (created by plot_KM)}
}
\value{
An object of class \code{quick_PH} containing the following elements:
\item{cloglogplot}{A log cumulative hazard plot in a \code{easysurv} theme}
\item{coxph_HR}{A \code{summary.coxPH} object containing statistical test
outputs, namely the Cox PH hazard ratio and confidence intervals.}
\item{coxzph_ph_test}{A \code{summary.coxPH} object containing statistical
test outputs}
\item{schoenfeld_plot}{A Schoenfeld residuals plot in a \code{easysurv}
theme}
}
\description{
Assess the proportional hazards assumptions with plots and statistical tests.
}
\examples{
\dontrun{
# Load required packages
library(survival)
library(easysurv)

input_data <- survival::lung

surv_data <- tibble(
  time = input_data$time,
  event = input_data$status - 1,
  strata = as.factor(input_data$sex)
)

PH_check <- quick_PH(
  data = surv_data,
  time = "time",
  event = "event",
  strata = "strata"
)
}
}
