% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_start_select.R
\name{quick_start_select}
\alias{quick_start_select}
\title{Launch an Example Script to Start Your Survival Analysis}
\usage{
quick_start_select(output_file_name = NULL, template_file_name = NULL)
}
\arguments{
\item{output_file_name}{Optional. A file name to use for the script. Defaults
to "easysurv_start.R".}

\item{template_file_name}{Optional. The name of the R script to be sourced
as a template. Defaults to "quick_template_lung.R".}
}
\description{
This function launches an example script for starting survival analysis
using the easysurv package. The code is inspired by usethis::use_template
but modified to work outside the context of an .RProj or package.
}
\examples{
\dontrun{
quick_start_select()
}
}
