% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_to_XL.R
\name{quick_to_XL}
\alias{quick_to_XL}
\title{Export \code{quick_...} output to Excel via \code{openxlsx}}
\usage{
quick_to_XL(wb, quick_object)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet}

\item{quick_object}{The output of quick_... command}
}
\description{
Export \code{quick_...} output to Excel via \code{openxlsx}
}
\examples{
\dontrun{
# To add
}
}
