% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMLE.cure.R
\name{runMLE.cure}
\alias{runMLE.cure}
\title{Run maximum likelihood estimation (MLE) on cure models for survival analysis}
\usage{
runMLE.cure(x, exArgs)
}
\arguments{
\item{x}{The specified distribution to run MLE on.
If \code{"splines"}, a flexible parametric survival model is fitted using
the \code{\link[flexsurv]{flexsurvspline}} function. If any other
distribution, a cure model is fitted using the
\code{\link[flexsurvcure]{flexsurvcure}} function.}

\item{exArgs}{Additional arguments to pass to the
\code{\link[flexsurvcure]{flexsurvcure}} or
\code{\link[flexsurv]{flexsurvspline}} function, such as the formula and
data.}
}
\value{
A list containing the fitted model, AIC, BIC, DIC (NULL in this
case), the time taken to run the estimation, and the model name.
}
\description{
This function performs MLE on cure models using either the
\code{\link[flexsurvcure]{flexsurvcure}} or
\code{\link[flexsurv]{flexsurvspline}} functions depending on the
specified distribution.
}
\examples{
\dontrun{
library(survival)

# Example usage with flexsurvcure
formula <- Surv(time, status) ~ as.factor(sex)
data <- lung
runMLE.cure("weibull", exArgs = list(formula = formula, data = data))

# Example usage with flexsurvspline
formula <- Surv(time, status) ~ as.factor(sex)
data <- lung
k <- 3
runMLE.cure("splines", exArgs = list(formula = formula, data = data, k = k))
}
}
