% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ph.R
\name{test_ph}
\alias{test_ph}
\title{Test Proportional Hazards Assumption}
\usage{
test_ph(data, time, event, group, plot_theme = ggplot2::theme_bw())
}
\arguments{
\item{data}{A data frame containing the survival data.}

\item{time}{The name of the column in \code{data} containing the
time-to-event information.}

\item{event}{The name of the column in \code{data} indicating whether the
event of interest occurred.}

\item{group}{The name of the column in \code{data} defining the grouping
variable.}

\item{plot_theme}{The theme to be used for the plots.}
}
\value{
A list containing plots and test results related to the assessment
of the proportional hazards assumption.
}
\description{
Assesses the proportional hazards assumption for survival data using a
Cox proportional hazards model and related tests.
}
\examples{
\dontrun{

ph_results <- get_ph(
  data = easysurv::easy_bc,
  time = "recyrs",
  event = "censrec",
  group = "group"
)
}
}
