% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_easysurv}
\alias{theme_easysurv}
\title{Plot theme for easysurv}
\usage{
theme_easysurv()
}
\description{
Plot theme for easysurv
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(ggsurvfit)
fit <- survfit2(Surv(time, status) ~ surg, data = df_colon)
fit |> ggsurvfit() + theme_easysurv()
\dontshow{\}) # examplesIf}
}
