% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easysurv-package.R
\docType{package}
\name{easysurv-package}
\alias{easysurv}
\alias{easysurv-package}
\title{easysurv: Tools to Simplify Survival Data Analysis and Model Fitting}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The easysurv package aims to simplify fitting parametric models for extrapolating survival data. It facilitates plotting Kaplan-Meier curves, assessing the proportional hazards assumption, estimating parametric survival models using engines such as flexsurv, flexsurvspline, flexsurvcure, and survival, and exporting the analyses to Excel.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Maple-Health-Group/easysurv}
  \item \url{https://maple-health-group.github.io/easysurv/}
  \item Report bugs at \url{https://github.com/Maple-Health-Group/easysurv/issues}
}

}
\author{
\strong{Maintainer}: Niall Davison \email{niall.davison@maplehealthgroup.com} (\href{https://orcid.org/0000-0002-9447-0825}{ORCID})

Authors:
\itemize{
  \item Brad Kievit \email{brad.kievit@maplehealthgroup.com}
}

}
\keyword{internal}
