% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easysurv-package.R
\docType{package}
\name{easysurv-package}
\alias{easysurv-package}
\alias{_PACKAGE}
\alias{easysurv}
\title{easysurv: Estimate and inspect parametric survival models}
\description{
Functions to streamline the process of fitting parametric models to extrapolate survival data. easysurv can fit data via simple parametric modelling, mixture cure modelling and spline modelling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Maple-Health-Group/easysurv}
  \item Report bugs at \url{https://github.com/Maple-Health-Group/easysurv/issues}
}

}
\author{
\strong{Maintainer}: Niall Davison \email{niall.davison@maplehealthgroup.com}

Authors:
\itemize{
  \item Brad Kievit \email{brad.kievit@maplehealthgroup.com}
}

}
\keyword{package}
