% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easysurv-package.R
\docType{package}
\name{easysurv-package}
\alias{easysurv}
\alias{easysurv-package}
\title{easysurv: Simplify Survival Data Analysis and Model Fitting}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Inspect survival data, plot Kaplan-Meier curves, assess the proportional hazards assumption, fit parametric survival models, predict and plot survival and hazards, and export the outputs to Excel. A simple interface for fitting survival models using the 'flexsurv', 'flexsurvspline', 'flexsurvcure', and 'survival' packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Maple-Health-Group/easysurv}
  \item \url{https://maple-health-group.github.io/easysurv/}
  \item Report bugs at \url{https://github.com/Maple-Health-Group/easysurv/issues}
}

}
\author{
\strong{Maintainer}: Niall Davison \email{niall.davison@maplehealthgroup.com} (\href{https://orcid.org/0000-0002-9447-0825}{ORCID})

Authors:
\itemize{
  \item Brad Kievit \email{brad.kievit@maplehealthgroup.com}
}

Other contributors:
\itemize{
  \item Maple Health Group, LLC [copyright holder, funder]
}

}
\keyword{internal}
