% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easysurv-package.R
\docType{package}
\name{easysurv-package}
\alias{easysurv}
\alias{easysurv-package}
\title{easysurv: Estimate and Inspect Parametric Survival Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions to streamline the process of fitting parametric models to extrapolate survival data. easysurv can fit data via simple parametric modelling, mixture cure modelling and spline modelling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Maple-Health-Group/easysurv}
  \item \url{https://maple-health-group.github.io/easysurv/}
  \item Report bugs at \url{https://github.com/Maple-Health-Group/easysurv/issues}
}

}
\author{
\strong{Maintainer}: Niall Davison \email{niall.davison@maplehealthgroup.com}

Authors:
\itemize{
  \item Brad Kievit \email{brad.kievit@maplehealthgroup.com}
}

}
\keyword{internal}
