% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_KM.R
\name{get_KM}
\alias{get_KM}
\title{Generate Kaplan-Meier estimates}
\usage{
get_KM(data, time, event, group = NULL, add_time_0 = FALSE, ...)
}
\arguments{
\item{data}{A data frame containing the survival data.}

\item{time}{The name of the column in \code{data} containing the
time-to-event information.}

\item{event}{The name of the column in \code{data} indicating whether the
event of interest occurred.}

\item{group}{(Optional) The name of the column in \code{data} defining the
grouping variable. Default is \code{NULL}.}

\item{add_time_0}{(Optional) Logical indicating whether to add time 0
to the Kaplan-Meier estimates. Default is \code{FALSE}.}

\item{...}{(Optional) Parameters to pass to ggsurvplot.}
}
\value{
A list containing Kaplan-Meier estimates, summary statistics, plots,
and additional outputs.
}
\description{
Calculates Kaplan-Meier estimates for survival data and returns summary
statistics, plots, and additional outputs.
}
\examples{
\dontrun{

KM_results <- get_KM(
 data = easysurv::easy_bc,
 time = "recyrs",
 event = "censrec",
 group = "group")

}
}
