% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fits}
\alias{plot_fits}
\title{Plot Survival Fits}
\usage{
plot_fits(pred_data, km_data, km_include = TRUE, title = NULL, subtitle = NULL)
}
\arguments{
\item{pred_data}{A data frame containing the survival predictions.}

\item{km_data}{A data frame containing the Kaplan-Meier survival data.}

\item{km_include}{Logical value indicating whether to include lines for
Kaplan-Meier survival data. Default is \code{TRUE}.}

\item{title}{Title of the plot. Default is \code{NULL}.}

\item{subtitle}{Subtitle of the plot. Default is \code{NULL}.}
}
\value{
A ggplot2 object.
}
\description{
Plots the survival fits generated by the \code{fit_models} function.
}
