% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_KM}
\alias{plot_KM}
\title{Plot Kaplan-Meier Data}
\usage{
plot_KM(
  fit,
  risk_table = TRUE,
  median_line = TRUE,
  legend_position = "bottom",
  plot_theme = ggplot2::theme_bw(),
  xlab = "Time",
  ylab = "Survival Probability (\%)"
)
}
\arguments{
\item{fit}{A \code{\link[survival]{survfit}} object representing the
survival data.}

\item{risk_table}{Logical value indicating whether to include a risk table
below the plot. Default is \code{TRUE}.}

\item{median_line}{Logical value indicating whether to include a line
representing the median survival time. Default is \code{TRUE}.}

\item{legend_position}{Position of the legend in the plot. Default is
"bottom".}

\item{plot_theme}{ggplot2 theme for the plot. Default is
\code{ggplot2::theme_bw()}.}

\item{xlab}{Label for the x-axis. Default is "Time".}

\item{ylab}{Label for the y-axis. Default is "Survival Probability (\%)".}
}
\value{
A ggplot object representing the Kaplan-Meier survival curve plot.
}
\description{
Generates a Kaplan-Meier survival curve plot using
\code{\link[ggsurvfit]{ggsurvfit}} with customisable options.
This function provides sensible defaults while allowing for customisation.
}
