% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_schoenfeld.R
\name{plot_schoenfeld}
\alias{plot_schoenfeld}
\title{Modified version of \code{\link[survminer]{ggcoxdiagnostics}} to replace
\code{gather_} and set \code{geom_smooth} formulae}
\usage{
plot_schoenfeld(
  fit_coxph,
  data,
  formula,
  hline = TRUE,
  sline = TRUE,
  sline.se = TRUE,
  hline.col = "red",
  hline.size = 1,
  hline.alpha = 1,
  hline.yintercept = 0,
  hline.lty = "dashed",
  sline.col = "blue",
  sline.size = 1,
  sline.alpha = 0.3,
  sline.lty = "dashed",
  point.col = "black",
  point.size = 1,
  point.shape = 19,
  point.alpha = 1,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  plot.theme = theme_bw()
)
}
\arguments{
\item{fit_coxph}{An object of class \code{\link[survival]{coxph.object}} -
created with \code{\link[survival]{coxph}} function.}

\item{data}{Data for fit_coxph}

\item{formula}{Formula for fit_coxph}

\item{hline}{a logical - should the horizontal line be added to highlight
the \code{Y=0} level.}

\item{sline, sline.se}{a logical - should the smooth line be added to
highlight the local average for residuals.}

\item{hline.col, hline.size, hline.lty, hline.alpha, hline.yintercept}{color,
size, linetype, visibility and Y-axis coordinate to be used for
\code{\link[ggplot2]{geom_hline}}.
Used only when \code{hline = TRUE}.}

\item{sline.col, sline.size, sline.lty, sline.alpha}{color, size,
linetype and visibility to be used for \code{\link[ggplot2]{geom_smooth}}.
Used only when \code{sline = TRUE}.}

\item{point.col, point.size, point.shape, point.alpha}{color, size, shape and
visibility to be used for points.}

\item{title, subtitle, caption}{main title, subtitle and caption.}

\item{plot.theme}{ggplot2 theme for the plot. Default is theme_bw()}
}
\value{
Returns an object of class \code{ggplot}.
}
\description{
Displays diagnostics graphs presenting goodness of Cox
Proportional Hazards Model fit, that
can be calculated with \code{\link[survival]{coxph}} function.
This function is largely identical to \code{ggcoxdiagnostics}
from the \code{survminer} package with some minor alterations in order to
accommodate \code{easysurv} functionality.
}
\examples{
\dontrun{
#to add
}

}
