% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_and_plot.R
\name{predict_and_plot}
\alias{predict_and_plot}
\title{Predict and plot \code{fit_models} on \code{data} at \code{eval_time}.}
\usage{
predict_and_plot(fit_models, eval_time = NULL, data, interval = "none")
}
\arguments{
\item{fit_models}{An object returned from fit_models.}

\item{eval_time}{(Optional) A vector of evaluation time points for generating
predictions. Default is \code{NULL}, which if left as NULL, generates a
sequence from 0 to 5 times the maximum observed time.}

\item{data}{The survival data used to estimate the models.}

\item{interval}{A character string specifying the type of interval to
predict. Options are "none" or "confidence". Default is "none".}
}
\description{
Predict and plot \code{fit_models} on \code{data} at \code{eval_time}.
}
