% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_start2.R
\name{quick_start2}
\alias{quick_start2}
\title{Launch an Example Script to Start Your Survival Analysis}
\usage{
quick_start2(output_file_name = NULL)
}
\arguments{
\item{output_file_name}{Optional. A file name to use for the script. Defaults
to "easysurv_start.R" within a helper function.}
}
\description{
This function launches an example script for starting survival analysis
using the easysurv package. The script uses the bc data set exported from
the flexsurv package. The code is inspired by usethis::use_template but
modified to work outside the context of an .RProj or package.
}
\examples{
\dontrun{
quick_start2()
}
}
