% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_starts.R
\name{quick_start3}
\alias{quick_start3}
\title{Launch Example Survival Analysis Script using the easy_adtte Data Set}
\usage{
quick_start3(output_file_name = NULL)
}
\arguments{
\item{output_file_name}{Optional. A file name to use for the script. Defaults
to "easysurv_start.R"  within a helper function.}
}
\value{
A new R script file with example code.
}
\description{
This function launches an example script for starting survival analysis
using the easysurv package. The script uses simulated phase III breast
cancer trial data available from the ggsurvfit package.
The code is inspired by \code{\link[usethis:use_template]{usethis::use_template()}}
but modified to work outside the context of an .RProj or package.
}
\examples{
quick_start3()
}
