% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMLE.cure.R
\name{runMLE.cure}
\alias{runMLE.cure}
\title{Run maximum likelihood estimation (MLE) on cure models for survival analysis}
\usage{
runMLE.cure(x, cure_weights, exArgs)
}
\arguments{
\item{x}{The specified distribution to run MLE on.}

\item{cure_weights}{Case weights. The function expects a string
corresponding to a variable name within the data. Can be NULL}

\item{exArgs}{Additional arguments to pass to the
\code{\link[flexsurvcure]{flexsurvcure}} function.}
}
\value{
A list containing the fitted model, AIC, BIC, DIC (NULL in this
case), the time taken to run the estimation, and the model name.
}
\description{
This function performs MLE on cure models using
\code{\link[flexsurvcure]{flexsurvcure}}
}
\examples{
\dontrun{
library(survival)

# Example usage with flexsurvcure
formula <- Surv(time, status) ~ as.factor(sex)
data <- lung
runMLE.cure("weibull", exArgs = list(formula = formula, data = data))
}
}
