% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_PH.R
\name{test_PH}
\alias{test_PH}
\title{Test Proportional Hazards Assumption}
\usage{
test_PH(
  data,
  time,
  event,
  group,
  plot.theme = theme_bw(),
  risk.table.theme = theme_bw()
)
}
\arguments{
\item{data}{A data frame containing the survival data.}

\item{time}{The name of the column in \code{data} containing the
time-to-event information.}

\item{event}{The name of the column in \code{data} indicating whether the
event of interest occurred.}

\item{group}{The name of the column in \code{data} defining the grouping
variable.}

\item{plot.theme}{(Optional) The theme for the ggplot2 plots. Default is
\code{theme_bw()}.}

\item{risk.table.theme}{(Optional) The theme for the risk table plot. Default
is \code{theme_bw()}.}
}
\value{
A list containing plots and test results related to the assessment
of the proportional hazards assumption.
}
\description{
Assesses the proportional hazards assumption for survival data using a
Cox proportional hazards model and related tests.
}
\examples{
\dontrun{

PH_results <- get_PH(
 data = easysurv::easy_bc,
 time = "recyrs",
 event = "censrec",
 group = "group")

}
}
