% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_easysurv}
\alias{theme_easysurv}
\title{Plot Theme for easysurv Survival and Hazard Plots}
\usage{
theme_easysurv()
}
\value{
A ggplot2 theme object.
}
\description{
Plot Theme for easysurv Survival and Hazard Plots
}
\examples{
library(ggsurvfit)
fit <- survfit2(Surv(time, status) ~ surg, data = df_colon)
fit |> ggsurvfit() + theme_easysurv()
}
