% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_risktable_easysurv}
\alias{theme_risktable_easysurv}
\title{Plot theme for easysurv risk tables}
\usage{
theme_risktable_easysurv()
}
\description{
To be used with \code{ggsurvfit::add_risktable()}.
}
