% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_xl.R
\name{write_to_xl}
\alias{write_to_xl}
\title{Export easysurv output to Excel via \code{openxlsx}}
\usage{
write_to_xl(wb, object)
}
\arguments{
\item{wb}{A Workbook object containing a worksheet}

\item{object}{The output of an easysurv command}
}
\description{
Export easysurv output to Excel via \code{openxlsx}
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

km_results <- get_km(
  data = easysurv::easy_bc,
  time = "recyrs",
  event = "censrec",
  group = "group"
)

wb <- openxlsx::createWorkbook()
write_to_xl(wb, km_results)

openxlsx::saveWorkbook(wb, "km_results.xlsx", overwrite = TRUE)
openxlsx::openXL("km_results.xlsx")
\dontshow{\}) # examplesIf}
}
